backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.0.1"
description = "Backend for laudio vision."
authors = ["Laudio Transformers <transformers@laudio.com>"]

# README file(s) are used as the package description
readme = ["README.md", "LICENSE.md"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.8"
alembic = "^1.9.2"
python-dotenv = "^0.21.1"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.5"
flask-restful = "^0.3.9"

[tool.poetry.group.dev.dependencies]
# Dependency groups are supported for organizing your dependencies
pytest = "^7.2.1"
black = "^22.12.0"
pylint = "^2.15.10"

[tool.poetry.scripts]
backend-start = 'src.__main__:main'

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
